Remove dependency on ostruct
Requiring ostruct on Ruby 3.4 warns:
```
/home/user/.rbenv/versions/3.4.1/lib/ruby/gems/3.4.0/gems/gem-release-2.2.2/lib/gem/release/data.rb:2: warning: /home/user/.rbenv/versions/3.4.1/lib/ruby/3.4.0/ostruct.rb was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.5.0.
You can add ostruct to your Gemfile or gemspec to silence this warning.
Also please contact the author of gem-release-2.2.2 to request adding ostruct into its gemspec.
```

But since the values that this ostruct will take is known, there is no need to use ostruct all all. I've simply replaced it with a normal struct.
It's a bit akward because of support for such old ruby versions (no `keyword_init`) but works out.